# packages/server/src/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Boardie",
    "description": "API to serve the Boardie application"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/me/friends/requests": {
      "post": {
        "summary": "Sends a friendship request to another user",
        "tags": [
          "friends"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userEmail": {
                    "type": "string",
                    "format": "email"
                  },
                  "userId": {
                    "type": "number"
                  }
                },
                "additionalProperties": false,
                "description": "Data used to send a friendship request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request was sent"
          },
          "404": {
            "description": "The requested user does not exist"
          },
          "409": {
            "description": "The friendship already exists or the request was already sent"
          }
        }
      },
      "get": {
        "summary": "Gets all friendship requests for the logged user",
        "tags": [
          "friends"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The logged user's friendship requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "userId": {
                        "type": "number"
                      },
                      "fullName": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "userId",
                      "fullName"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/me/friends/requests/:requestingUserId": {
      "put": {
        "summary": "Answers a friendship request",
        "tags": [
          "friends"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "requestingUserId",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accept": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "accept"
                ],
                "additionalProperties": false,
                "description": "Data used to accept a friendship request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request answer has been processed"
          },
          "404": {
            "description": "The friendship request does not exist"
          },
          "409": {
            "description": "The friendship already exists"
          }
        }
      }
    },
    "/me/friends/:friendUserId": {
      "delete": {
        "summary": "Deletes a friendship",
        "tags": [
          "friends"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "friendUserId",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The friendship was deleted"
          },
          "404": {
            "description": "The friendship does not exist"
          }
        }
      }
    },
...
  }
}